+++ OSPaymentsLib/Protocols/OSPMTRequestDelegate.swift
@@ -25,8 +25,19 @@ protocol OSPMTRequestDelegate: AnyObject {
     /// Sets Payment details and triggers its processing.
     /// - Parameters:
     ///   - detailsModel: payment details information.
+    ///   - accessToken: Authorisation token related with a full payment type.
     ///   - completion: an async closure that can return a successful Payment Scope Model or an error otherwise.
-    func trigger(with detailsModel: OSPMTDetailsModel, _ completion: @escaping OSPMTCompletionHandler)
+    func trigger(with detailsModel: OSPMTDetailsModel, and accessToken: String?, _ completion: @escaping OSPMTCompletionHandler)
+}
+
+extension OSPMTRequestDelegate {
+    /// Sets Payment details and triggers its processing. It uses the default method without the `accessToken` parameter.
+    /// - Parameters:
+    ///   - detailsModel: payment details information.
+    ///   - completion: an async closure that can return a successful Payment Scope Model or an error otherwise.
+    func trigger(with detailsModel: OSPMTDetailsModel, _ completion: @escaping OSPMTCompletionHandler) {
+        self.trigger(with: detailsModel, and: nil, completion)
+    }
 }
 
 /// Class that implements the `OSPMTRequestDelegate` for Apple Pay, providing it the required that details to work.
@@ -36,6 +47,8 @@ class OSPMTApplePayRequestBehaviour: NSObject, OSPMTRequestDelegate {
     
     var paymentStatus: PKPaymentAuthorizationStatus = .failure
     var paymentScope: OSPMTScopeModel?
+    var paymentDetails: OSPMTDetailsModel?
+    var accessToken: String?
     var completionHandler: OSPMTCompletionHandler!
     
     /// Constructor method.
@@ -47,7 +60,14 @@ class OSPMTApplePayRequestBehaviour: NSObject, OSPMTRequestDelegate {
         self.requestTriggerType = requestTriggerType
     }
     
-    func trigger(with detailsModel: OSPMTDetailsModel, _ completion: @escaping OSPMTCompletionHandler) {
+    /// Sets Payment details and triggers its processing.
+    /// - Parameters:
+    ///   - detailsModel: payment details information.
+    ///   - accessToken: Authorisation token related with a full payment type.
+    ///   - completion: an async closure that can return a successful Payment Scope Model or an error otherwise.
+    func trigger(with detailsModel: OSPMTDetailsModel, and accessToken: String?, _ completion: @escaping OSPMTCompletionHandler) {
+        self.paymentDetails = detailsModel
+        self.accessToken = accessToken
         self.completionHandler = completion
         
         let result = self.requestTriggerType.createRequestTriggerBehaviour(for: detailsModel, andDelegate: self)
@@ -110,14 +130,46 @@ extension OSPMTApplePayRequestBehaviour: PKPaymentAuthorizationControllerDelegat
     ///   - payment: The authorized payment. This object contains the payment token you need to submit to your payment processor, as well as the billing and shipping information required by the payment request.
     ///   - completion: The completion handler to call with the result of authorizing the payment.
     func paymentAuthorizationController(_ controller: PKPaymentAuthorizationController, didAuthorizePayment payment: PKPayment, handler completion: @escaping (PKPaymentAuthorizationResult) -> Void) {
-        if let scopeModel = payment.createScopeModel() {
-            self.paymentScope = scopeModel
-            self.paymentStatus = .success
-        } else {
-            self.paymentScope = nil
-            self.paymentStatus = .failure
+        func setPaymentResults(with errorArray: [OSPMTError], and scopeModel: OSPMTScopeModel?, _ completion: @escaping (PKPaymentAuthorizationResult) -> Void) {
+            if errorArray.isEmpty, let scopeModel = scopeModel {
+                self.paymentScope = scopeModel
+                self.paymentStatus = .success
+            } else {
+                self.paymentScope = nil
+                self.paymentStatus = .failure
+            }
+            
+            completion(PKPaymentAuthorizationResult(status: self.paymentStatus, errors: errorArray))
         }
         
-        completion(PKPaymentAuthorizationResult(status: self.paymentStatus, errors: []))
+        if let paymentDetails = paymentDetails, paymentDetails.gateway != nil {
+            guard let accessToken = self.accessToken, !accessToken.isEmpty else {
+                return completion(PKPaymentAuthorizationResult(status: self.paymentStatus, errors: [OSPMTError.tokenIssue]))
+            }
+            
+            guard let paymentGateway = self.configuration.gatewayModel, paymentGateway.gatewayEnum == paymentDetails.gateway else {
+                return completion(PKPaymentAuthorizationResult(status: self.paymentStatus, errors: [OSPMTError.gatewayNotConfigured]))
+            }
+            
+            guard let gatewayWrapper = OSPMTGatewayFactory.createWrapper(for: paymentGateway) else {
+                return completion(PKPaymentAuthorizationResult(status: self.paymentStatus, errors: [OSPMTError.gatewaySetFailed]))
+            }
+            
+            gatewayWrapper.process(payment, with: paymentDetails, and: accessToken) { result in
+                var errorArray = [OSPMTError]()
+                var paymentResultModel: OSPMTServiceProviderInfoModel?
+                
+                switch result {
+                case .success(let result):
+                    paymentResultModel = result
+                case .failure(let error):
+                    errorArray += [error]
+                }
+                
+                setPaymentResults(with: errorArray, and: payment.createScopeModel(for: paymentResultModel), completion)
+            }
+        } else {
+            setPaymentResults(with: [], and: payment.createScopeModel(), completion)
+        }
     }
 }