+++ docs/CHANGELOG.md
@@ -4,7 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## 1.1.0
+
+### 2023-01-05
+- Feat: Add access token to Full Payment Process (https://outsystemsrd.atlassian.net/browse/RMET-2147).
+
+### 2022-12-12
+- Feat: Add Payment Service Provider property to `OSPMTDetailsModel` struct (https://outsystemsrd.atlassian.net/browse/RMET-2095).
+
+### 2022-12-02
+- Feat: Add Stripe as Payment Service Provider (https://outsystemsrd.atlassian.net/browse/RMET-2078).
 
 ### 2022-11-14
 - Add `podspec` file.